FEAT: Add very ugly placeholder for scryptsy
