+++ external/CMakeLists.txt
@@ -4,7 +4,7 @@ file(GLOB external
         bech32/bech32.h bech32/bech32.cpp
         cliparser/cxxopts.hpp
         toml/cpptoml.h
-        aes_128_ctr/aes.hpp
+        aes_128_ctr/aes.c
         )
 
 add_library(external ${external})
